requirements.txt
influxdb_client==1.49.0
matplotlib==3.10.8
neo4j==5.15.0
networkx==3.6.1
numpy==2.4.0
pymongo==4.6.1
python_bcrypt==0.3.2
redis==5.0.1
